member_1_speech/requirements.txt
pyaudio==0.2.14
openai
python-dotenv
pvporcupine==3.0.1
pyautogui
numpy==1.26.4